Update Scribe imageSrc in listOfProjects.js
I've been updating the images for the repos and renamed the one that's referenced here. I think that doing the rounded icon would be best for the site anyway :)

Want to thank you again, btw! Looking at the traffic for Scribe, most of the external traffic is coming from firstcontributions.github.io 😊